requirements.txt
streamlit==1.32.0
langchain-groq==0.1.0
langchain-community==0.0.20
langchain-core==0.1.0
ollama==0.1.0
faiss-cpu==1.7.4
python-dotenv==1.0.0